Added start of server file and package dependencies
Index contains some server side code that should implement the socket.io connections on connect to the webpage. It should handle instances where a user selects "Login" or "Register". This should also create the MySQL databases if they are not already present. Some details will need to be added and adapted when the front-end page for the login comes in.